chore: rename module to rqmetrics
