+++ rqmetrics/__main__.py
@@ -1,5 +1,5 @@
 """Entry point for RQ's exporter."""
-from rqexport.cli import main
+from rqmetrics.cli import main
 
 if __name__ == "__main__":
     main()
